Update: Refactor logic function
