+++ src/features/translation/api/index.ts
@@ -1,0 +1,1 @@
+export { useTranslation } from "./use-translation";
